Update annual summary API view
Modified weekday_sums so that it is an annual total of all weekday ons and offs
Added saturday_sum so that it is an annual total of all saturday ons and offs
Added sunday_sum so that it is an annual total of all sunday ons and offs

annual_sums should equal weekday_sums + saturday_sums + sunday_sums

This is an untested change. Feel free to change this or redo it or whatever you need to do.